+++ CHANGES.md
@@ -1,6 +1,7 @@
 v0.8.1
 ------
 * Fix for infer on multiple JSON files to combine them before inference
+* Fix to handle nested Objects in inference
 
 v0.8.0
 ------